Split a couple of API calls out into a service
